refactor: constants to routes
